# docs/swagger.yaml
basePath: /
definitions:
  errs.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.ProductType:
    properties:
      ProdType_Id:
        type: integer
      ProdType_Name:
        type: string
    type: object
  models.ProductTypeCreate:
    properties:
      ProdType_Id:
        minimum: 0
        type: integer
      ProdType_Name:
        maxLength: 40
        type: string
    required:
    - ProdType_Id
    - ProdType_Name
    type: object
  models.ProductTypeUpdate:
    properties:
      ProdType_Name:
        maxLength: 40
        type: string
    required:
    - ProdType_Name
    type: object
  models.Response:
    properties:
      message:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: example@gmail.com
    name: Walter White
    url: https://twitter.com/example
  description: Gin-Test - Teletubbie's ProductType API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin-Test - ProductType API
  version: "1.0"
paths:
  /producttype/:
    get:
      consumes:
      - application/json
      description: Get all producttype
      produces:
      - application/json
      responses:
        "200":
          description: Get ProductTypes Successfully
          schema:
            items:
              $ref: '#/definitions/models.ProductType'
            type: array
        "404":
          description: Error Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Error Unexpected Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Get All ProductType
      tags:
      - producttype
    post:
      consumes:
      - application/json
      description: Create producttype
      parameters:
      - description: ProductType data to be create
        in: body
        name: ProductType
        required: true
        schema:
          $ref: '#/definitions/models.ProductTypeCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Create ProductType Successfully
          schema:
            $ref: '#/definitions/models.ProductType'
        "400":
          description: Error Bad Request
          schema:
            $ref: '#/definitions/errs.AppError'
        "409":
          description: Error Conflict Error
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Error Unexpected Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Create ProductType
      tags:
      - producttype
  /producttype/{id}:
    delete:
      consumes:
      - application/json
      description: Delete producttype by id
      parameters:
      - description: ProductType ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete ProductType Successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Error Bad Request
          schema:
            $ref: '#/definitions/errs.AppError'
        "404":
          description: Error Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Error Unexpected Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Delete ProductType
      tags:
      - producttype
    get:
      consumes:
      - application/json
      description: Get producttype by id
      parameters:
      - description: ProductType ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get ProductType Successfully
          schema:
            $ref: '#/definitions/models.ProductType'
        "400":
          description: Error Bad Request
          schema:
            $ref: '#/definitions/errs.AppError'
        "404":
          description: Error Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Error Unexpected Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Get ProductType
      tags:
      - producttype
    put:
      consumes:
      - application/json
      description: Update producttype by id
      parameters:
      - description: ProductType ID
        in: path
        name: id
        required: true
        type: integer
      - description: ProductType data to be update
        in: body
        name: ProductType
        required: true
        schema:
          $ref: '#/definitions/models.ProductTypeUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Update ProductType Successfully
          schema:
            $ref: '#/definitions/models.ProductType'
        "400":
          description: Error Bad Request
          schema:
            $ref: '#/definitions/errs.AppError'
        "404":
          description: Error Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Error Unexpected Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Update ProductType
      tags:
      - producttype
  /producttype/count:
    get:
      consumes:
      - application/json
      description: Get producttype's count from database
      produces:
      - application/json
      responses:
        "200":
          description: Get ProductType's Count Successfully
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Error Unexpected Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Get ProductType Count
      tags:
      - producttype
  /producttype/health:
    get:
      consumes:
      - application/json
      description: Health check
      operationId: HealthCheckHandler
      produces:
      - application/json
      responses:
        "200":
          description: ProductType service is running
          schema:
            $ref: '#/definitions/models.Response'
      summary: Health Check
      tags:
      - producttype
schemes:
- http
- https
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"